requirements.txt
streamlit>=1.28.0
pandas>=2.1.1
openpyxl>=3.1.2
plotly>=5.17.0
firebase-admin>=6.4.0
pyrebase4>=4.7.1
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.3
bcrypt>=4.0.1
